refactor: split html method into directives
