+++ src/index.ts
@@ -2,12 +2,12 @@ export { window } from './window';
 export { connect, disconnect } from './helpers';
 export { customElements } from './CustomElementRegistry';
 export { DOM } from './DOM';
-export { Fragment, html, h, render } from './render';
+export { Fragment, compile, html, h, render } from './render';
 export { css } from './css';
 export { listen, delegateEventListener, undelegateEventListener, dispatchEvent, dispatchAsyncEvent, defineListeners } from './events';
 export { property, state, observe, getProperty, getProperties, defineProperties, defineProperty } from './property';
 export { extend, Component, isComponent, isComponentConstructor, customElement } from './Component';
-export { until } from './directives';
+export { parseDOM, until } from './directives';
 
 export type { CustomElementRegistry } from './CustomElementRegistry';
 export type { Observable } from './Observable';
